deleted axios, connected react-scripts
