+++ .gitignore
@@ -2,3 +2,4 @@
 node_modules/
 coverage/
 __pycache__/
+.eslintcache
