--- .htaccess
@@ -1,2 +1,0 @@
-RewriteEngine On
-RewriteRule ^(\w+)$ ?page=$1
