--- index.html
@@ -1,13 +1,0 @@
-<html lang="ru">
-    <head>
-        <title>Rap-Paraphraser</title>
-        <script defer src="public/script.js"></script>
-        <link rel = "stylesheet" href="public/style.css" />
-        <link href="https://fonts.googleapis.com/css?family=Montserrat&display=swap" rel="stylesheet">
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-    </head>
-    <body>
-        <div id = "root"></div>
-    </body>
-</html>
