+++ package.json
@@ -1,40 +1,34 @@
 {
   "scripts": {
-    "build": "webpack -p",
-    "test": "jest",
-    "lint": "eslint src",
-    "watch": "webpack --watch --mode=development",
-    "server": "python -m http.server --cgi 8000"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "version": "1.2.8",
+  "version": "2.0.0",
   "dependencies": {
-    "axios": "^0.20.0",
     "classnames": "^2.2.6",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-router-dom": "^5.1.2"
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.1",
+    "web-vitals": "^0.2.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.10.4",
-    "@jest/globals": "^26.4.2",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^5.0.0",
-    "file-loader": "^6.1.1",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.5",
-    "eslint": "^7.11.0",
-    "eslint-plugin-react": "^7.21.5",
-    "jest-cli": "^26.4.2",
-    "mini-css-extract-plugin": "^1.1.1",
-    "node-sass": "^4.13.1",
-    "path": "^0.12.7",
-    "react-test-renderer": "^16.12.0",
-    "sass-loader": "^10.0.3",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.10"
+    "@testing-library/jest-dom": "^5.11.8",
+    "@testing-library/react": "^11.2.2",
+    "@testing-library/user-event": "^12.6.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
