--- public/style.css
@@ -1,2 +1,0 @@
-body,body form *{font-family:'Montserrat', sans-serif;font-size:16pt}body{width:60%;margin:10px auto auto}h1{font-size:26pt;text-align:center}.size14{font-size:14pt}#p1{text-align:justify}p.center{text-align:center}a,a:active,a:visited{color:#00e}a:hover{color:red}.main-div{display:flex}.main-div #div1,.main-div #div2{width:50%}.main-div #div1{padding-right:10px}.main-div #div2{padding-left:10px}.main-div .button-bar{text-align:center;width:100%;margin-top:10px}.main-div .button-bar button{margin:5px 5px}.main-div #enterWord{width:100%}.main-div #enterText{width:100%;height:200px}.main-div #result{border:1px solid black;height:200px;overflow-y:auto}.main-div #result.loading{background:url(loading.gif) no-repeat center}.main-div .ws-pre{white-space:pre}.main-div .ws-pre-wrap{white-space:pre-wrap}
-
