+++ src/components/Header.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import './css/Header.css';
+
+export const Header = () => {
+    return (
+        <>
+            <h1>Rap-Paraphraser</h1>
+            <p>
+                Добро пожаловать на сайт для перефразирования текста любого типа (в том числе и
+                одного слова) в семантическом поле русского репа с добавлением реп-сленга. Алгоритм
+                перефразирования основан на использовании нейронной сети Word2Vec, обученной на
+                текстах более, чем 9500 песен русских исполнителей, которая находит близкие по
+                семантике слова в контексте русского репа.
+            </p>
+        </>
+    );
+};
