+++ src/components/css/Header.css
@@ -1,0 +1,9 @@
+h1 {
+    font-size: 32px;
+    text-align: center;
+}
+
+p {
+    text-align: justify;
+    font-size: 20px;
+}
