+++ src/components/css/ResultField.css
@@ -1,0 +1,9 @@
+.result {
+    border: 1px solid black;
+    height: 200px;
+    overflow-y: auto;
+}
+
+.result.loading {
+     /*background: url("../../assets/img/loading.gif") no-repeat center;*/
+ }
