+++ src/index.css
@@ -1,0 +1,20 @@
+body {
+    width: 60%;
+    margin: 10px auto auto;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+    font-size: 16pt;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+}
+
+form * {
+    font-size: 16pt;
+}
+
+a, a:active, a:visited {
+    color: rgb(0, 0, 238);
+}
+
+a:hover {
+    color: red;
+}
