+++ src/pages/css/AssociationsPage.css
@@ -1,0 +1,19 @@
+.main-div { display: flex; font-size: 18px; }
+
+.main-div > div { width: 50% }
+.main-div > div:nth-of-type(1) { padding-right: 10px; }
+.main-div > div:nth-of-type(2) { padding-left: 10px; }
+
+.button-bar {
+    text-align: center;
+    width: 100%;
+    margin-top: 10px;
+}
+
+.button-bar button { margin: 5px 5px; }
+
+button, input { font-size: 18px; }
+
+.enter-word { width: 100%; }
+
+.ws-pre { white-space: pre; }
