+++ src/pages/css/MainPage.css
@@ -1,0 +1,1 @@
+.main-page p { text-align: center; }
