--- src/styles/index.scss
@@ -1,59 +1,0 @@
-%textFont {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 16pt;
-}
-
-body {
-  width: 60%;
-  margin: 10px auto auto;
-  @extend %textFont;
-
-  form * { @extend %textFont; }
-}
-
-h1 {
-  font-size: 26pt;
-  text-align: center;
-}
-
-.size14 { font-size: 14pt; }
-
-#p1 { text-align: justify; }
-
-p.center { text-align: center; }
-
-a, a:active, a:visited { color: rgb(0, 0, 238); }
-a:hover { color:red; }
-
-.main-div {
-  display: flex;
-
-  #div1, #div2 { width: 50%; }
-  #div1 { padding-right: 10px; }
-  #div2 { padding-left: 10px; }
-
-  .button-bar {
-    text-align: center;
-    width: 100%;
-    margin-top: 10px;
-
-    button { margin: 5px 5px; }
-  }
-
-  #enterWord { width: 100%; }
-
-  #enterText { width: 100%; height: 200px; }
-
-  #result {
-    border: 1px solid black;
-    height: 200px;
-    overflow-y: auto;
-
-    &.loading {
-      background: url("../../assets/img/loading.gif") no-repeat center;
-    }
-  }
-
-  .ws-pre { white-space: pre; }
-  .ws-pre-wrap { white-space: pre-wrap; }
-}
