--- webpack.config.js
@@ -1,45 +1,0 @@
-const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
-module.exports = {
-    entry: './src/js/index.js',
-    output: {
-        filename: './script.js',
-        path: path.resolve(__dirname, 'public')
-    },
-    plugins: [
-        new MiniCssExtractPlugin({
-            filename: './style.css'
-        })
-    ],
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader'
-                }
-            },
-            {
-                test: /\.(css|scss)$/,
-                exclude: /node_modules/,
-                use: [
-                    MiniCssExtractPlugin.loader,
-                    'css-loader', // translates CSS into CommonJS
-                    'sass-loader', // compiles Sass to CSS, using Node Sass by default
-                ]
-            },
-            {
-                test: /\.(png|jpg|gif)$/,
-                use: {
-                    loader: 'url-loader',
-                    options: {
-                        limit: 1024,
-                        name: '[name].[ext]'
-                    }
-                }
-            }
-        ]
-    },
-};
